refactor(not found): added not found pages to reusable component module
