+++ client/src/features/registration/RegistrationRoutes.tsx
@@ -2,12 +2,14 @@ import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Signin from './pages/Signin';
 import SignUp from './pages/SignUp';
+import NotFound from '../../common/components/NotFound';
 
 export default function RegistrationRoutes() {
   return (
     <Routes>
       <Route path="/" element={<Signin />} />
       <Route path="/signup" element={<SignUp />} />
+      <Route path="*" element={<NotFound />} />
     </Routes>
   );
 }
